Define caching for frontend files
This should ensure that browsers always get the newest index.html of the
SPA. The CSS and JS files can be cached quite agressively, because their
name is bound to their content. As long as the index.html is the current
one, we'll know which CSS and JS files to use.